backend/requirements.txt
fastapi>=0.115.5
uvicorn[standard]>=0.32.0
chronos-forecasting>=2.0.0
pandas>=2.2.0
openai>=1.52.2
SQLAlchemy>=2.0.32
asyncpg>=0.29.0
alembic>=1.13.2

